Refactored the new product form into some reusable components
